Separate Authorization into separate sections
