+++ pylintrc
@@ -4,7 +4,8 @@ max-line-length=120
 disable=duplicate-code,  # reports false alarms AND can't be disabled locally; pylint issue #214
         fixme,  # ignore TODOs
         redefined-outer-name,
-        too-many-arguments
+        too-many-arguments,
+        too-few-public-methods
 
 [BASIC]
 good-names=i,j,k,
